Skills list command has been extracted into separate file.
