+++ src/cmd_god/show.cpp
@@ -8,6 +8,7 @@
 #include "ban.h"
 #include "administration/privilege.h"
 #include "cmd/do_features.h"
+#include "cmd/do_skills.h"
 #include "cmd/do_spells.h"
 #include "cmd_god/shutdown_parameters.h"
 #include "communication/parcel.h"
@@ -37,7 +38,6 @@ extern const char *Dirs[];
 extern bool is_empty(ZoneRnum zone_nr);
 extern void show_apply(CharData *ch, CharData *vict);
 extern void print_rune_stats(CharData *ch);
-extern void list_skills(CharData *ch, CharData *vict, const char *filter = nullptr);
 
 void ShowClassInfo(CharData *ch, const std::string &class_name, const std::string &params) {
 	if (class_name.empty()) {
@@ -607,7 +607,7 @@ void do_show(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/) {
 				SendMsgToChar("Нет такого игрока.\r\n", ch);
 				return;
 			}
-			list_skills(vict, ch);
+			DisplaySkills(vict, ch);
 			break;
 		case 14:        // show spells
 			if (!*value) {
